+++ Modules/Project.Windows.ProgramInfo/Public/Test-Environment.ps1
@@ -63,6 +63,7 @@ function Test-Environment
 	}
 
 	# TODO: We need target computer system drive instead of localmachine systemdrive
+	# TODO: We need regex instead of wildcard here, ex. C:\\ or C:\ is both valid
 	if (($FilePath -like "$env:SystemDrive\Users\*") -or
 		[array]::Find($UserProfileEnvironment, [System.Predicate[string]] { $FilePath -like "$($args[0])*" }))
 	{